DJANGO_API/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==4.2.29
djangorestframework==3.17.0
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
